+++ Stopwatch/AppInfo.h
@@ -2,7 +2,7 @@
 
 // ReSharper disable IdentifierTypo
 
-#define VER_PRODUCTNAME_STR         "SimpleTimer"
+#define VER_PRODUCTNAME_STR         "Stopwatch"
 #define VER_FILE_DESCRIPTION_STR    "Control named out-of-process Timers"
 
 #include "../Common/AppInfo.h"
