+++ Stopwatch/NamedTimerRepository.h
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-namespace SimpleTimer
+namespace Stopwatch
 {
     class NamedTimerRepository
     {
