+++ Stopwatch/stdafx.cpp
